BackEnd/requirements.txt
django
pillow
rest_framework_simplejwt
djangorestframework-simplejwt
django-cors-headers